+++ .gitignore
@@ -127,3 +127,8 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# Temp bmp file
+temp.bmp
+
+
